requirements.txt
celery==5.2.7
fastapi==0.85.0
Pillow==9.2.0
python-multipart==0.0.5
redis==4.4.0
uvicorn==0.18.3

# dev
pytest==7.2.0